Implements namespaces for better optimization
